requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
Markdown==3.5.1
python-frontmatter==1.0.0
PyYAML==6.0.1
nltk==3.8.1
textstat==0.7.3
jieba==0.42.1
python-dotenv==1.0.0
gunicorn==21.2.0 